jfactivexx: MEx merge
Conflicts:
	arch/arm/mach-msm/devices-8064.c
	arch/arm/mach-msm/hotplug.c
	arch/arm/mach-msm/pm-8x60.c
	drivers/input/keyboard/gpio_keys.c
	drivers/media/video/msm/sensors/jc_v4l2.c
	drivers/mfd/max77693.c
	drivers/video/msm/mdp4_overlay.c
	drivers/video/msm/vidc/common/vcd/vcd_power_sm.c

+++ drivers/media/video/msm/sensors/msm_sensor.h
@@ -41,9 +41,9 @@ void msm_sensor_group_hold_off(struct msm_sensor_ctrl_t *s_ctrl);
 int32_t msm_sensor_set_fps(struct msm_sensor_ctrl_t *s_ctrl,
 			struct fps_cfg   *fps);
 int32_t msm_sensor_write_exp_gain1(struct msm_sensor_ctrl_t *s_ctrl,
-		uint16_t gain, uint32_t line);
+		uint32_t gain, uint32_t line);
 int32_t msm_sensor_write_exp_gain2(struct msm_sensor_ctrl_t *s_ctrl,
-		uint16_t gain, uint32_t line);
+		uint32_t gain, uint32_t line);
 int32_t msm_sensor_set_sensor_mode(struct msm_sensor_ctrl_t *s_ctrl,
 	int mode, int res);
 int32_t msm_sensor_mode_init(struct msm_sensor_ctrl_t *s_ctrl,
